Move Braze components to package
