+++ packages/braze/browser/index.js
@@ -1,0 +1,5 @@
+const PreferenceCenter = () => import(/* webpackChunkName: "braze-preference-center" */ './preference-center.vue');
+
+export default (Browser) => {
+  Browser.register('BrazePreferenceCenter', PreferenceCenter);
+};
