+++ packages/braze/browser/subscription-group.vue
@@ -10,13 +10,11 @@
         class="custom-control-input"
       >
       <label class="custom-control-label" :for="`custom-boolean-${id}`" :required="required">
-        <strong v-html="label" />
+        <strong>{{ label }}</strong>
         <strong v-if="required" class="text-danger">*</strong>
-        <span
-          v-if="description"
-          class="font-weight-normal d-block text-muted"
-          v-html="description"
-        />
+        <span v-if="description" class="font-weight-normal d-block text-muted">
+          {{ description }}
+        </span>
       </label>
     </div>
   </div>
