+++ packages/braze/env.js
@@ -1,0 +1,8 @@
+const { cleanEnv, validators } = require('@parameter1/base-cms-env');
+
+const { nonemptystr } = validators;
+
+module.exports = cleanEnv(process.env, {
+  RECAPTCHA_V3_SITE_KEY: nonemptystr({ desc: 'The Recaptcha Site key.' }),
+  RECAPTCHA_V3_SECRET_KEY: nonemptystr({ desc: 'The Recaptcha secret key.' }),
+});
