+++ packages/braze/graphql/mutations/idx-user-update-receive-email.js
@@ -1,0 +1,11 @@
+const gql = require('graphql-tag');
+
+module.exports = gql`
+mutation SetAppUserReceiveEmail($input: UpdateAppUserMutationInput!) {
+  updateAppUser(input: $input) {
+    id
+    email
+    receiveEmail
+  }
+}
+`;
