+++ packages/braze/package.json
@@ -12,8 +12,15 @@
     "test": "yarn compile && yarn lint"
   },
   "dependencies": {
+    "@parameter1/base-cms-env": "^3.0.0",
+    "@parameter1/base-cms-marko-web": "^3.8.0",
+    "@parameter1/base-cms-marko-web-gtm": "^3.7.3",
+    "@parameter1/base-cms-marko-web-recaptcha": "^3.0.0",
     "@parameter1/base-cms-object-path": "^3.0.0",
+    "@parameter1/base-cms-utils": "^3.0.0",
     "@parameter1/joi": "^1.2.10",
+    "express": "^4.17.1",
+    "graphql-tag": "^2.10.1",
     "node-fetch": "^2.6.6"
   }
 }
