+++ packages/braze/routes/index.js
@@ -1,0 +1,5 @@
+const preferenceCenter = require('./preference-center');
+
+module.exports = (app) => {
+  preferenceCenter(app);
+};
