+++ packages/global/browser/index.js
@@ -1,9 +1,9 @@
 
 import MonoRail from '@parameter1/base-cms-marko-web-theme-monorail/browser';
+import Braze from '@science-medicine-group/package-braze/browser';
 import FormLogin from './form-login.vue';
 
 const ContentMeterTrack = () => import(/* webpackChunkName: "content-meter-tracker" */ './track-content-meter.vue');
-const PreferenceCenter = () => import(/* webpackChunkName: "braze-preference-center" */ './braze-preference-center.vue');
 
 export default (Browser) => {
   MonoRail(Browser, {
@@ -13,5 +13,5 @@ export default (Browser) => {
     },
   });
   Browser.register('ContentMeterTrack', ContentMeterTrack);
-  Browser.register('PreferenceCenter', PreferenceCenter);
+  Braze(Browser);
 };
