+++ packages/global/package.json
@@ -10,7 +10,6 @@
     "test": "yarn lint"
   },
   "dependencies": {
-    "@parameter1/base-cms-env": "^3.0.0",
     "@parameter1/base-cms-image": "^3.8.0",
     "@parameter1/base-cms-inflector": "^3.0.0",
     "@parameter1/base-cms-marko-core": "^3.3.1",
@@ -27,7 +26,6 @@
     "@parameter1/base-cms-marko-web-omeda": "^3.7.3",
     "@parameter1/base-cms-marko-web-omeda-identity-x": "^3.10.0",
     "@parameter1/base-cms-marko-web-p1-events": "^3.7.1",
-    "@parameter1/base-cms-marko-web-recaptcha": "^3.0.0",
     "@parameter1/base-cms-marko-web-reveal-ad": "^3.0.0",
     "@parameter1/base-cms-marko-web-search": "^3.3.1",
     "@parameter1/base-cms-marko-web-social-sharing": "^3.0.0",
