+++ packages/global/routes/index.js
@@ -2,7 +2,7 @@ const htmlSitemap = require('@parameter1/base-cms-marko-web-html-sitemap/routes'
 const renderBlock = require('@parameter1/base-cms-marko-web-theme-monorail/routes/render-block');
 const search = require('@parameter1/base-cms-marko-web-theme-monorail/routes/search');
 const taxonomy = require('@parameter1/base-cms-marko-web-theme-monorail/routes/taxonomy');
-const braze = require('./braze');
+const braze = require('@science-medicine-group/package-braze/routes');
 const feed = require('./feed');
 const nativeX = require('./native-x');
 const printContent = require('./print-content');
